refactor: change from match to challenge
